Contined work on materials system, renderer, etc Removed some deprecated files Graphics portion of engine is in a mess right now until a renderer/scene graph system is properly integrated and tested

+++ SORE_lib/sore_profiler.h
@@ -41,14 +41,14 @@ namespace SORE_Profiler
 					avgTime = 0.0;
 					intervalCount = 0;
 				}
-				unsigned int   startTime;
-				unsigned int   endTime;
+				double         startTime;
+				double         endTime;
 				unsigned int   timesRun;
-				unsigned int   lastTime;
+				double         lastTime;
 				
-				         float avgTime;
-								 float shortAvgTime;
-				         int   intervalCount;
+				double         avgTime;
+				double         shortAvgTime;
+				int            intervalCount;
 				unsigned int   minTime;
 				unsigned int   maxTime;
 				std::string name;
